Added matrixes, netchannelinfo, engine interface. Added material and PlayerInfo struct with their regarded enums and structs.

+++ source-sdk/utils/constants/const.hpp
@@ -2,7 +2,7 @@
 #include <cstdint>
 
 namespace utils::consts {
-	constexpr size_t pointerSize = sizeof(void*);
+	constexpr uint8_t pointerSize = sizeof(void*);
 	constexpr const char* lastLoadedModule = "serverbrowser.dll";
 	constexpr const char* prefix = "meowware > ";
 }
